Conflicts:
	protected/views/student/keyExer.php

+++ protected/controllers/AdminController.php
@@ -62,8 +62,18 @@ class AdminController extends CController {
             $condition = $condition . "''";
             $rows = Student::model()->deleteAll("userID in ($condition)");
         }
+         if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM teacher WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $pages = $array_stuLst['pages'];
         $stuLst = Student::model()->findAll("is_delete = '1'");
         $this->render('recycleStu', array(
+            'pages' => $pages,
             'stuLst' => $stuLst,
             'rows' => $rows
         ));
@@ -98,8 +108,19 @@ class AdminController extends CController {
                 'is_delete' => '0'
                     ), "userID in ($condition)");
         }
+         if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM student WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $pages = $array_stuLst['pages'];
+        
         $stuLst = Student::model()->findAll("is_delete = '1'");
         $this->render('recycleStu', array(
+            'pages' => $pages,
             'stuLst' => $stuLst,
             'rows' => $rows
         ));
@@ -117,9 +138,19 @@ class AdminController extends CController {
     }
 
     public function actionRecycleStu() {
-        $stuLst = Student::model()->findAll("is_delete = '1'");
+        if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM student WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $stuLst = $array_stuLst['list'];
+        $pages = $array_stuLst['pages'];
         $this->render('recycleStu', array(
-            'stuLst' => $stuLst
+            'stuLst' => $stuLst,
+            'pages' => $pages
         ));
     }
 
@@ -1104,8 +1135,19 @@ class AdminController extends CController {
             TeacherClass::model()->deleteAll("teacherID in ($condition)");
             $rows = Teacher::model()->deleteAll("userID in ($condition)");
         }
+         if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM teacher WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $teaLst = $array_stuLst['list'];
+        $pages = $array_stuLst['pages'];
         $teaLst = Teacher::model()->findAll("is_delete = '1'");
         $this->render('recycleTea', array(
+            'pages' => $pages,
             'teaLst' => $teaLst,
             'rows' => $rows
         ));
@@ -1140,18 +1182,39 @@ class AdminController extends CController {
                 'is_delete' => '0'
                     ), "userID in ($condition)");
         }
+        if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM student WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $pages = $array_stuLst['pages'];
         $teaLst = Teacher::model()->findAll("is_delete = '1'");
         $this->render('recycleTea', array(
+            'pages'=>$pages,
             'teaLst' => $teaLst,
             'rows' => $rows
         ));
     }
 
     public function actionRecycleTea() {
-        $teaLst = Teacher::model()->findAll("is_delete = '1'");
+         if(isset($_GET ['page'])){
+            Yii::app()->session ['lastPage'] = $_GET ['page'];
+        }else {
+            Yii::app()->session ['lastPage'] = 1;
+        }
+         Yii::app()->session ['lastUrl'] = "recyclestu";
+        $sql = "SELECT * FROM teacher WHERE is_delete = '1'";
+        $array_stuLst = Tool::pager($sql,10);
+        $teaLst = $array_stuLst['list'];
+        $pages = $array_stuLst['pages'];
         $this->render('recycleTea', array(
-            'teaLst' => $teaLst
+            'teaLst' => $teaLst,
+            'pages' => $pages
         ));
+        
     }
 
     public function actionClassLst() {
@@ -1274,12 +1337,15 @@ class AdminController extends CController {
 
     public function actionAddClass() {
         $result = 'no';
+         $sql = "SELECT * FROM `tb_class` ORDER BY `tb_class`.`classID` DESC ";
+        $aClass = Yii::app()->db->createCommand($sql)->query();
         if (isset($_POST ['className'])) {
             $className = $_POST ['className'];
             $classes = TbClass::model()->findAll("className = '$className'");
             if (count($classes) > 0) {
                 $result = 2;
             } else {
+                Yii::app()->session['insert_class']=$className;
                 $classID = TbClass::model()->insertClass($_POST ['className'], $_POST ['courseID']);
                 $lessons = Lesson::model()->findall('classID=? and courseID=?', array(0, $_POST ['courseID']));
                 foreach ($lessons as $lesson) {
@@ -1289,7 +1355,8 @@ class AdminController extends CController {
             }
         }
         $this->render('addClass', [
-            'result' => $result
+            'result' => $result,
+            'allClass' => $aClass
         ]);
     }
 
@@ -1343,14 +1410,16 @@ class AdminController extends CController {
         $curLesson = $class ['currentLesson'];
 
         $sql = "SELECT * FROM student WHERE classID = '$classID' AND is_delete = 0";
-        $criteria = new CDbCriteria ();
-        $stus = Yii::app()->db->createCommand($sql)->query();
+        $array_stuLst = Tool::pager($sql,4);
+        $stus = $array_stuLst['list'];
         $nums = $stus->rowCount;
-
+        $pages_stu = $array_stuLst['pages'];
+        
         $sql = "SELECT * FROM teacher_class WHERE classID =$classID";
         $teacherOfClass = Yii::app()->db->createCommand($sql)->query();
 
         $this->render('infoCLass', array(
+            'pages_stu' =>$pages_stu,
             'classID' => $classID,
             'className' => $className,
             'curCourse' => $curCourse,
@@ -1365,10 +1434,14 @@ class AdminController extends CController {
 
     public function actionAddStuClass() {
         $sql = "SELECT * FROM student WHERE classID = '0' AND is_delete = 0";
-        $result = Yii::app()->db->createCommand($sql)->query();
+        $array_class = Tool::pager($sql, 10);
+        $result = $array_class['list'];
+        $pages = $array_class['pages'];
+        
         $this->render('addStuClass', array(
             'classID' => $_GET ["classID"],
-            'posts' => $result
+            'posts' => $result,
+            'pages' =>$pages,
                 )
         );
     }
@@ -1376,11 +1449,16 @@ class AdminController extends CController {
     public function actionAddTeaClass() {
         $classID = $_GET ["classID"];
         $sql = "SELECT teacherID FROM teacher_class WHERE classID = '$classID'  order by teacherID ASC";
+        $sqltea="SELECT * FROM teacher order by userID ASC";
+        $array_tea = Tool::pager($sqltea, 10);
+        $teacher = $array_tea['list'];
+        $pages = $array_tea['pages'];
         $result = Yii::app()->db->createCommand($sql)->query();
         $this->render('addTeaClass', array(
+            'pages'=>$pages,
             'classID' => $classID,
             'posts' => $result,
-            'teachers' => TbClass::model()->teaInClass()
+            'teachers' => $teacher,
                 )
         );
     }
@@ -1724,9 +1802,12 @@ class AdminController extends CController {
             $i = 2;
             $answer = $_POST ['in1'];
             for (; $i <= 3 * 10; $i ++) {
-                if ($_POST ['in' . $i] != "")
-                    $answer = $answer . "$" . $_POST ['in' . $i];
-                else
+                if ($_POST ['in' . $i] != ""){
+                   if ($i % 3 == 0)
+                        $answer = $answer . "_" . $_POST['in' . $i];
+                    else
+                        $answer = $answer . ":" . $_POST['in' . $i];
+                }else
                     break;
             }
             $result = KeyType::model()->insertKey($_POST ['title'], $answer, 0);
@@ -1741,6 +1822,7 @@ class AdminController extends CController {
         $sql = "SELECT * FROM key_type WHERE exerciseID = '$exerciseID'";
         $result = Yii::app()->db->createCommand($sql)->query();
         $result = $result->read();
+        $result['content'] = str_replace("_", ":", $result['content']);
         if (!isset($_GET ['action'])) {
             $this->render("editKey", array(
                 'exerciseID' => $exerciseID,
@@ -1765,7 +1847,7 @@ class AdminController extends CController {
         $answer = $_POST ['in1'];
         for (; $i <= 3 * 10; $i ++) {
             if ($_POST ['in' . $i] != "")
-                $answer = $answer . "$" . $_POST ['in' . $i];
+                $answer = $answer . ":" . $_POST ['in' . $i];
             else
                 break;
         }
@@ -2703,7 +2785,7 @@ class AdminController extends CController {
         foreach ($courseLst_forNumber as $v){
             $number = 0;
             $courseID = $v['courseID'];
-            $lesson = Lesson::model()->findAll("courseID = '$courseID'");
+            $lesson = Lesson::model()->findAll("courseID = '$courseID' and classID = 0");
             if(empty($lesson)){
             }else{
                 foreach ($lesson as $value){
@@ -2762,11 +2844,26 @@ class AdminController extends CController {
 
     public function actionAddCourse() {
         $result = 'no';
+        $sql = "SELECT * FROM `course` ORDER BY `course`.`courseID` DESC ";
+        $aCourse = Yii::app()->db->createCommand($sql)->query();
         if (isset($_POST ['courseName'])) {
+            $flag = 1;
             $courseNumber = $_POST['courseNumber'];
             $courseName = $_POST['courseName'];
-            $result = Course::model()->insertCourse($_POST ['courseName'], 0);
+            $allCourse = Course::model()->findAll();
+            foreach ($allCourse as $v){
+                if($v['courseName']==$courseName){
+                    $result = 'have_same_course';
+                    $flag = 0;
+                    break;
+                }
+                $flag = 1;
+            }
+            if($flag==1){
+                 $result = Course::model()->insertCourse($_POST ['courseName'], 0);
+            }
             if ($result == 1) {
+                Yii::app()->session['insert_course'] = $_POST ['courseName'];
                 $sql = "SELECT MAX(courseID) AS id FROM course";
                 $max_id = Yii::app()->db->createCommand($sql)->query()->read();
                 $courseID = $max_id['id'];
@@ -2782,13 +2879,15 @@ class AdminController extends CController {
                         }
                     }
                 }
-                $this->render('addCourse', [
-                    'result' => $result
-                ]);
             }
+            $this->render('addCourse', [
+                    'result' => $result,
+                    'allCourse' =>$aCourse
+                ]);
         } else {
             $this->render('addCourse', [
-                'result' => $result
+                'result' => $result,
+                'allCourse' =>$aCourse
             ]);
         }
     }
@@ -2841,7 +2940,10 @@ class AdminController extends CController {
         $courseName = $_GET ['courseName'];
         $createPerson = $_GET ['createPerson'];
         $result = 'no';
+        $sql = "SELECT * FROM `lesson` WHERE `courseID` = $courseID AND `classID` = 0 ORDER BY `lesson`.`number` DESC ";
+        $aLesson = Yii::app()->db->createCommand($sql)->query();
         if (isset($_POST['lessonName'])) {
+            Yii::app()->session['insert_lesson'] = $_POST['lessonName'];
             $result = Lesson::model()->insertLesson($_POST['lessonName'], $courseID, 0, 0);
             $classes = TbClass::model()->findall("currentCourse = '$courseID'");
             foreach ($classes as $class) {
@@ -2852,7 +2954,8 @@ class AdminController extends CController {
             'courseID' => $courseID,
             'courseName' => $courseName,
             'createPerson' => $createPerson,
-            'result' => $result
+            'result' => $result,
+            'allLesson' =>$aLesson
         ));
     }
 
@@ -2880,18 +2983,14 @@ class AdminController extends CController {
             return;
         }
         if ($_FILES["file"]["type"] == "application/vnd.ms-powerpoint") {
-            if ($_FILES["file"]["size"] < 30000000) {
-                if ($_FILES["file"]["error"] > 0) {
-                    $result = "Return Code: " . $_FILES["file"]["error"];
-                } else {
-                    $newName = Tool::createID() . ".ppt";
-                    $oldName = $_FILES["file"]["name"];
-                    move_uploaded_file($_FILES["file"]["tmp_name"], $dir . iconv("UTF-8", "gb2312", $newName));
-                    Resourse::model()->insertRela($newName, $oldName);
-                    $result = "上传成功！";
-                }
+            if ($_FILES["file"]["error"] > 0) {
+                $result = "Return Code: " . $_FILES["file"]["error"];
             } else {
-                $reult = "PPT文件限定大小为30M！";
+                $newName = Tool::createID() . ".ppt";
+                $oldName = $_FILES["file"]["name"];
+                move_uploaded_file($_FILES["file"]["tmp_name"], $dir . iconv("UTF-8", "gb2312", $newName));
+                Resourse::model()->insertRela($newName, $oldName);
+                $result = "上传成功！";
             }
         } else {
             $result = "请上传正确类型的文件！";
@@ -2972,18 +3071,14 @@ class AdminController extends CController {
             return;
         }
         if ($_FILES["file"]["type"] == "video/mp4" || $_FILES["file"]["type"] == "application/octet-stream") {
-            if ($_FILES["file"]["size"] < 200000000) {
-                if ($_FILES["file"]["error"] > 0) {
-                    $result = "Return Code: " . $_FILES["file"]["error"];
-                } else {
-                    $oldName = $_FILES["file"]["name"];
-                    $newName = Tool::createID() . "." . pathinfo($oldName, PATHINFO_EXTENSION);
-                    move_uploaded_file($_FILES["file"]["tmp_name"], $dir . iconv("UTF-8", "gb2312", $newName));
-                    Resourse::model()->insertRela($newName, $oldName);
-                    $result = "上传成功！";
-                }
+            if ($_FILES["file"]["error"] > 0) {
+                $result = "Return Code: " . $_FILES["file"]["error"];
             } else {
-                $reult = "视频文件限定大小为200M！";
+                $oldName = $_FILES["file"]["name"];
+                $newName = Tool::createID() . "." . pathinfo($oldName, PATHINFO_EXTENSION);
+                move_uploaded_file($_FILES["file"]["tmp_name"], $dir . iconv("UTF-8", "gb2312", $newName));
+                Resourse::model()->insertRela($newName, $oldName);
+                $result = "上传成功！";
             }
         } else {
             $result = "请上传正确类型的文件！";
@@ -3019,18 +3114,31 @@ class AdminController extends CController {
    }
    public function ActionDeleteNotice()
      {
+       if (isset($_POST['checkbox'])) {
+            $userIDlist = $_POST['checkbox'];
+            foreach ($userIDlist as $v) {
+                Notice::model()->delNotice($v);
+            }
+            $result = Notice::model()->findNotice();
+            $noticeRecord = $result ['noticeLst'];
+            $pages = $result ['pages'];
+            $this->render('noticeLst',  array('noticeRecord'=>$noticeRecord,'pages'=>$pages));
+        }
          $id = $_GET['id'];
          Notice::model()->deleteAll("id='$id'");
-         
          $result=Notice::model()->findNotice();
         $noticeRecord=$result ['noticeLst'];
         $pages = $result ['pages'];
        $this->render('noticeLst',  array('noticeRecord'=>$noticeRecord,'pages'=>$pages));
      }
      public function ActionNoticeContent(){
+         $result=0;
+        if(isset($_GET['action'])&&$_GET['action']=='edit'){
+            $result=1;
+        }
        $id = $_GET['id'];
        $noticeRecord=Notice::model()->find("id= '$id'");
-       $this->render('noticeContent',  array('noticeRecord'=>$noticeRecord));
+       $this->render('noticeContent',  array('noticeRecord'=>$noticeRecord,'result'=>$result));
      }
 
     public function actionSchedule() {
@@ -3062,7 +3170,11 @@ class AdminController extends CController {
         }
 
 
-        $class = TbClass::model()->findAll("");
+        $classsql = "SELECT * FROM tb_class";
+        $array_classLst = Tool::pager($classsql,10);
+        $class = $array_classLst['list'];
+        $classPages = $array_classLst['pages'];
+        
         $class_course = array();
         foreach ($class as $v) {
             $courseID = $v['currentCourse'];
@@ -3070,8 +3182,12 @@ class AdminController extends CController {
             $class_v = array("classID" => $v['classID'], "className" => $v['className'], "currentCourse" => $v['currentCourse'], "currentLesson" => $v['currentLesson'], "courseName" => $sqlCourse['courseName']);
             array_push($class_course, $class_v);
         }
-        $teacher = Teacher::model()->findAll("");
-        return $this->render('schedule', ['teacher' => $teacher, 'class' => $class_course]);
+        $teasql = "SELECT * FROM teacher WHERE is_delete = '0'";
+        $array_teaLst = Tool::pager($teasql,10);
+        $teaLst = $array_teaLst['list'];
+        $teaPages = $array_teaLst['pages'];
+        
+        return $this->render('schedule', ['class_pages'=>$classPages,'tea_pages'=>$teaPages,'teacher' => $teaLst, 'class' => $class_course]);
     }
 
     public function actionScheduleDetil() {
